App/requirements.txt
seaborn
numpy
pandas
streamlit
plotly.express
scikit-learn
imblearn